pyproject.toml

[project]
name = "hexamed-asset-management"
version = "0.1.0"
description = "Hexamed Asset Management System"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "xlrd>=2.0.2",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["app", "main", "models", "routes", "migrate_db", "setup_tables"]
